Get the new messages on the page to re-render!
The problem was that I need to not use props. I could, but it would be annoying to maintain.
Now we have a weird bug where after we submit, the page gets the same message  multiple times (has to do with subscription being called multiple times to change the cached result of the query given to the component -- over and over again).