Now the site displays the message LAST instead of FIRST.
Order matters in the array, and things with ... don't necessarily have to go last, depending on the context!

For more info on whatever the ... operator does, see
https://stackoverflow.com/a/5072906. I like this answer a lot because as the chad put it, "This is a
great answer because of all the clear examples for each use case. Thank
you for taking the time to write all this out."